api small fix

+++ backend/api/helpers/get_datafromApi.py
@@ -9,4 +9,4 @@ def getSimplePricedata(ticker:str,daysback=30,column="Open")->list:
     stock_data = yf.download(ticker,start=(curr_date-timedelta(daysback)),end=curr_date)
     stock_data = pd.DataFrame(stock_data)
     stock_data[column] = np.round(stock_data[column],3)
-    return pd.array(stock_data[column])
+    return pd.array(stock_data[column]).astype(float)
